moved modbus to the app.connection in an effort to reduce serial errors
