+++ PMAN_Server_Flask/website/pman_blueprints/kamoer_dip_1500/__init__.py
@@ -1,7 +1,7 @@
 from flask import current_app, request, Blueprint
 import json
 import pdb
-from .modbus import ModbusRTU, uint32_to_bytes, extract_data_cl, extract_data_hr
+from website.modbus import extract_data_cl
 from math import log, floor
 from ..utils import extract_pman_args 
 from serial import SerialException
@@ -25,8 +25,7 @@ def start_pump(rpm, direction, addr):
     addr = addr.to_bytes(1, byteorder='big')
     direction = int(direction)
     rpm = float(rpm)
-    modbus = ModbusRTU(current_app.connection.serial)
-    modbus.set_rpm(rpm, addr)
+    modbus = current_app.connection.modbus
     if direction == 1:
         modbus.motor_clockwise(addr)
     else:
@@ -39,8 +38,7 @@ def start_pump(rpm, direction, addr):
 @extract_pman_args
 def stop_pump(addr):
     addr = addr.to_bytes(1, byteorder='big')
-    modbus = ModbusRTU(current_app.connection.serial)
-    modbus.motor_stop(addr)
+    current_app.connection.modbus.motor_stop(addr)
     return {'status': "ok", 'message': "Pump Stopped"}
 
 @kamoer_peri.get('/status')
@@ -48,7 +46,7 @@ def read_status():
     addr_str = request.args.get('addr')  
     addr_int = int(addr_str) 
     addr_bytes = addr_int.to_bytes(1, byteorder='big') 
-    modbus = ModbusRTU(current_app.connection.serial)
+    modbus = current_app.connection.modbus
     data = modbus.read_motor_status(addr_bytes) # error here? contains b''
     data2 = modbus.read_dir(addr_bytes)
     if(data != b'' and data2 != b''):
