+++ PMAN_Server_Flask/website/templates/kamoer-peri/control.html
@@ -95,7 +95,7 @@ setInterval(() => {
 		})
 		.catch(error => console.error('Error:', error));
 	})
-}, 1000);
+}, 3000);
 
 {% for addr in addrs %}
 enforceCharset("rpm-{{addr}}","0123456789")
